# docs/swagger-openapi.yml
openapi: 3.0.1
info:
  title: Tài liệu REST API Nexora-Commerce
  description: Tài liệu RESTful API hiệu năng cao cho hệ thống thương mại điện tử
    Nexora-Commerce phục vụ tích hợp frontend.
  contact:
    name: Nexora Core Dev Team
    email: dev@nexoracommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Máy chủ phát triển cục bộ (Local Development Server)
security:
- bearerAuth: []
...
paths:
...
  /api/v1/brands/{brandId}:
    put:
      tags:
      - Thương hiệu
      summary: Cập nhật thương hiệu (Chỉ Admin)
      description: Yêu cầu quyền ADMIN. Thay đổi các thuộc tính của thương hiệu bao
        gồm tên, logo hoặc đường dẫn slug.
      operationId: updateBrand
      parameters:
      - name: brandId
        in: path
        description: ID thương hiệu trong cơ sở dữ liệu
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandRequest'
        required: true
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '200':
          description: Cập nhật thương hiệu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '404':
          description: Không tìm thấy thương hiệu
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '400':
          description: Dữ liệu đầu vào không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
      security:
      - bearerAuth: []
    delete:
      tags:
      - Thương hiệu
      summary: Xóa thương hiệu (Chỉ Admin)
      description: Yêu cầu quyền ADMIN. Xóa vĩnh viễn thương hiệu khỏi hệ thống.
      operationId: deleteBrand
      parameters:
      - name: brandId
        in: path
        description: ID thương hiệu trong cơ sở dữ liệu
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '200':
          description: Xóa thương hiệu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '404':
          description: Không tìm thấy thương hiệu
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
      - bearerAuth: []
...
  /api/v1/brands:
    get:
      tags:
      - Thương hiệu
      summary: Lấy danh sách tất cả thương hiệu
      description: Truy xuất danh sách tất cả các thương hiệu sản phẩm đang hoạt động.
      operationId: getAllBrands
      responses:
        '200':
          description: Lấy danh sách thương hiệu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListBrandResponse'
    post:
      tags:
      - Thương hiệu
      summary: Tạo thương hiệu mới (Chỉ Admin)
      description: Yêu cầu quyền ADMIN. Thêm mới một thương hiệu sản phẩm vào hệ thống.
      operationId: createBrand
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandRequest'
        required: true
      responses:
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '201':
          description: Tạo thương hiệu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '400':
          description: Dữ liệu đầu vào không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
      security:
      - bearerAuth: []
...
  /api/v1/brands/{idOrSlug}:
    get:
      tags:
      - Thương hiệu
      summary: Lấy chi tiết thương hiệu theo ID hoặc Slug
      description: Lấy chi tiết thương hiệu thông qua ID số trong cơ sở dữ liệu hoặc
        thông qua chuỗi slug của thương hiệu.
      operationId: getBrand
      parameters:
      - name: idOrSlug
        in: path
        description: ID thương hiệu hoặc chuỗi slug thương hiệu
        required: true
        schema:
          type: string
        example: apple
      responses:
        '404':
          description: Không tìm thấy thương hiệu
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
        '200':
          description: Lấy chi tiết thương hiệu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseBrandResponse'
...
  /api/v1/admin/brands:
    get:
      tags:
      - Quản lý Thương hiệu (Admin)
      summary: Lấy toàn bộ thương hiệu (Admin)
      description: Yêu cầu quyền ADMIN. Lấy danh sách phân trang tất cả thương hiệu
        sản phẩm phục vụ quản lý thương hiệu.
      operationId: getAllBrands_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Lấy danh sách thương hiệu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListBrandResponse'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    BrandRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 0
          type: string
        slug:
          maxLength: 100
          minLength: 0
          type: string
    ApiResponseBrandResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/BrandResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    BrandResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    ApiResponseListBrandResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/BrandResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Please paste your JWT Access Token in the box below to authorize.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT